dom/media/tests/mochitest/webvtt/HTMLTrackElement_track.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
	<title>Test for Bug 833386 - HTMLTrackElement</title>
	<script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
	<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
 	<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
	<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=833386">Mozilla Bug { 833386 } - HTMLTrackElement - Default </a><br>
	<video width="320" height="240" controls>
		<track id="subtitlesEN" src="captions/fast.vtt" srclang="en"/>
	</video>
	<pre id="test">
	<script class="testbody" type="text/javascript">
	/** Test Function that verifies whether a texttrackcuelist has the methods addCue and removeCue **/
	function test1(){
		var tracks = document.getElementById("subtitlesEN");
		is(tracks.track, tracks.track, "Test 1 - Expecting an object return");
		is(tracks.track.addCue, "function addCue() { [native code] }", "Test 1 - Object should have 'addCue' property");
		is(tracks.track.removeCue, "function removeCue() { [native code] }", "Test 1 - Object Should have 'removeCue' property");
	}
	/** Test for Bug {833386} **/
	SimpleTest.waitForExplicitFinish();
	// Verify WebVTT support
	var WebvttEnabled = undefined;
	try {
	  WebvttEnabled = SpecialPowers.getBoolPref("media.webvtt.enabled");
	} catch (ex) {
	// SpecialPowers failed, perhaps because WebVTT isn't compiled in
	console.log("media.webvtt.enabled pref not found; skipping WebVTT validation");
	}
	if (WebvttEnabled !== undefined) {
	   SpecialPowers.setBoolPref("media.webvtt.enabled", true);
	   addLoadEvent(test1);
	   SpecialPowers.setBoolPref("media.webvtt.enabled", false);
	   addLoadEvent(test1);
	   SpecialPowers.setBoolPref("media.webvtt.enabled", WebvttEnabled);
	}
	SimpleTest.finish();
	
	
	</script>
	</pre>		
</body>
</html>
